add check_and_move in archive

+++ src/utils_check_size.c
@@ -2,7 +2,7 @@
 
 void	*check_and_move(t_bin_file *file, void *dest, uint64_t needed_size)
 {
-	if ((size_t)((dest + needed_size) - file->ptr) > file->size)
+	if ((uint64_t)(dest + needed_size - file->ptr) > file->size)
 		return (NULL);
 	return (dest);
 }
